some 64bit DEC Alpha patches

+++ l3bitstream.c
@@ -2,9 +2,12 @@
  * ISO MPEG Audio Subgroup Software Simulation Group (1996)
  * ISO 13818-3 MPEG-2 Audio Encoder - Lower Sampling Frequency Extension
  *
- * $Id: l3bitstream.c,v 1.13 2000/02/19 13:32:30 afaber Exp $
+ * $Id: l3bitstream.c,v 1.14 2000/02/21 23:05:05 markt Exp $
  *
  * $Log: l3bitstream.c,v $
+ * Revision 1.14  2000/02/21 23:05:05  markt
+ * some 64bit DEC Alpha patches
+ *
  * Revision 1.13  2000/02/19 13:32:30  afaber
  * Fixed many warning messages when compiling with MSVC
  *
@@ -721,9 +724,9 @@ Huffmancodebits( BF_PartHolder **pph, int *ix, gr_info *gi )
 	  tables, we will pad with ones
 	*/
 	while ( stuffingWords-- )
-	    *pph = BF_addEntry( *pph, ~0, 32 );
+	    *pph = BF_addEntry( *pph, ~(u_int)0, 32 );
 	if ( remainingBits )
-	    *pph = BF_addEntry( *pph, ~0, remainingBits );
+	    *pph = BF_addEntry( *pph, ~(u_int)0, remainingBits );
 	bitsWritten += stuffingBits;
     }
     assert( bitsWritten == (int)(gi->part2_3_length - gi->part2_length) );
